+++ workspace/solutions/ecmascript/src/s0/s9/s6/s2/s_0962_00.js
@@ -1,0 +1,35 @@
+/*
+ * @lc app=leetcode.cn id=962 lang=javascript
+ *
+ * [962] 最大宽度坡
+ */
+
+// @lc code=start
+/**
+ * @param {number[]} nums
+ * @return {number}
+ */
+const maxWidthRamp = function (nums) {
+    const n = nums.length;
+    const stack = [];
+
+    for (let i = 0; i < n; i++) {
+        if (stack.length === 0 || nums[stack.at(-1)] > nums[i]) {
+            stack.push(i);
+        }
+    }
+
+    let result = 0;
+    for (let j = n - 1; j >= 0; j--) {
+        while (stack.length > 0 && nums[stack.at(-1)] <= nums[j]) {
+            result = Math.max(result, j - stack.pop());
+        }
+        if (stack.length === 0) {
+            break;
+        }
+    }
+    return result;
+};
+// @lc code=end
+
+export default maxWidthRamp;