+++ workspace/solutions/rust/src/s0/s9/mod.rs
@@ -1,2 +1,3 @@
 pub mod s0;
 pub mod s1;
+pub mod s6;