+++ workspace/solutions/rust/src/s0/s9/s6/s2/mod.rs
@@ -1,0 +1,5 @@
+pub mod s_0962_00;
+pub mod s_0962_01;
+pub mod s_0962_test;
+
+pub type TSolution = fn(Vec<i32>) -> i32;