+++ workspace/solutions/rust/src/s0/s9/s6/s2/s_0962_test.rs
@@ -1,0 +1,22 @@
+#[cfg(test)]
+mod tests {
+    use super::super::s_0962_00;
+    use super::super::s_0962_01;
+    use crate::utils::{json::ValueExtension, test::t};
+
+    #[test]
+    fn test() -> () {
+        t(
+            module_path!(), //
+            &vec![
+                s_0962_00::SOLUTION, //
+                s_0962_01::SOLUTION,
+            ],
+            |solve, input| {
+                serde_json::Value::from(solve(
+                    input[0].to_i32_array().unwrap(), //
+                ))
+            },
+        );
+    }
+}